feat(plugin): Migrate datasource dialect and enum implementations to plugin-default-impl  - Migrate DatabaseDialect implementations (5 files) using git mv  - Migrate Enum classes (5 files) using git mv  - Migrate related tests (19 files) using git mv  - Update DatabaseDialectManager to use SPI-loaded dialects  - Update SPI service file paths  - Update import statements in mapper implementations  - Preserve git history for all migrated files
