+++ plugin-default-impl/nacos-default-datasource-plugin/src/main/java/com/alibaba/nacos/plugin/datasource/impl/enums/derby/TrustedDerbylFunctionEnum.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 1999-2022 Alibaba Group Holding Ltd.
+ * Copyright 1999-2018 Alibaba Group Holding Ltd.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,33 +14,30 @@
  * limitations under the License.
  */
 
-package com.alibaba.nacos.plugin.datasource.enums.derby;
+package com.alibaba.nacos.plugin.datasource.impl.enums.derby;
 
 import java.util.HashMap;
 import java.util.Map;
 
 /**
- * Trusted Derby Function Enum.
+ * The TrustedSqlFunctionEnum enum class is used to enumerate and manage a list of trusted built-in SQL functions.
+ * By using this enum, you can verify whether a given SQL function is part of the trusted functions list
+ * to avoid potential SQL injection risks.
  *
- * @author xiweng.yy
+ * @author blake.qiu
  */
-public enum TrustedDerbyFunctionEnum {
+public enum TrustedDerbylFunctionEnum {
 
     /**
      * NOW().
      */
-    NOW("NOW()", "CURRENT_TIMESTAMP"),
+    NOW("NOW()", "CURRENT_TIMESTAMP");
 
-    /**
-     * CURRENT_TIMESTAMP().
-     */
-    CURRENT_TIMESTAMP("CURRENT_TIMESTAMP()", "CURRENT_TIMESTAMP");
-
-    private static final Map<String, String> LOOKUP_MAP = new HashMap<>();
+    private static final Map<String, TrustedDerbylFunctionEnum> LOOKUP_MAP = new HashMap<>();
 
     static {
-        for (TrustedDerbyFunctionEnum entry : TrustedDerbyFunctionEnum.values()) {
-            LOOKUP_MAP.put(entry.functionName, entry.function);
+        for (TrustedDerbylFunctionEnum entry : TrustedDerbylFunctionEnum.values()) {
+            LOOKUP_MAP.put(entry.functionName, entry);
         }
     }
 
@@ -48,7 +45,7 @@ public enum TrustedDerbyFunctionEnum {
 
     private final String function;
 
-    TrustedDerbyFunctionEnum(String functionName, String function) {
+    TrustedDerbylFunctionEnum(String functionName, String function) {
         this.functionName = functionName;
         this.function = function;
     }
@@ -60,9 +57,9 @@ public enum TrustedDerbyFunctionEnum {
      * @return function
      */
     public static String getFunctionByName(String functionName) {
-        String function = LOOKUP_MAP.get(functionName);
-        if (null != function) {
-            return function;
+        TrustedDerbylFunctionEnum entry = LOOKUP_MAP.get(functionName);
+        if (entry != null) {
+            return entry.function;
         }
         throw new IllegalArgumentException(String.format("Invalid function name: %s", functionName));
     }
