--- plugin/datasource/src/main/java/com/alibaba/nacos/plugin/datasource/dialect/MysqlDatabaseDialect.java
@@ -1,38 +1,0 @@
-/*
- * Copyright 1999-2022 Alibaba Group Holding Ltd.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.alibaba.nacos.plugin.datasource.dialect;
-
-import com.alibaba.nacos.plugin.datasource.constants.DatabaseTypeConstant;
-import com.alibaba.nacos.plugin.datasource.enums.mysql.TrustedMysqlFunctionEnum;
-
-/**
- * MySQL database dialect.
- *
- * @author xiweng.yy
- */
-public class MysqlDatabaseDialect extends AbstractDatabaseDialect {
-
-    @Override
-    public String getType() {
-        return DatabaseTypeConstant.MYSQL;
-    }
-
-    @Override
-    public String getFunction(String functionName) {
-        return TrustedMysqlFunctionEnum.getFunctionByName(functionName);
-    }
-}
