+++ plugin/datasource/src/test/java/com/alibaba/nacos/plugin/datasource/proxy/MapperProxyTest.java
@@ -16,7 +16,6 @@
 
 package com.alibaba.nacos.plugin.datasource.proxy;
 
-import com.alibaba.nacos.plugin.datasource.enums.mysql.TrustedMysqlFunctionEnum;
 import com.alibaba.nacos.plugin.datasource.mapper.Mapper;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -79,7 +78,7 @@ class MapperProxyTest {
             
             @Override
             public String getFunction(String functionName) {
-                return TrustedMysqlFunctionEnum.getFunctionByName(functionName);
+                return functionName;
             }
         };
         Mapper proxy = mapperProxy.createProxy(mapper);
